reusable types added to custom types for admin pages
