+++ Frontend/src/components/Admin/AdminOrders/AdminOrders.tsx
@@ -1,42 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-import { URL } from '../../../utils/URL';
+import { URL } from '../../../utils/URL.ts';
 import { IoIosArrowDown, IoIosArrowUp } from 'react-icons/io';
-import {Address} from '../../../custom.Types/userTypes.ts';
-
-export type ProductId = {
-  quantity: number;
-  productId: string;
-  image: string;
-  productName: string;
-  description: string;
-  price: number;
-  category: {
-    categoryName: string;
-  };
-};
-export type Product = {
-  quantity: number;
-  productId: ProductId;
-};
-export type Orders = {
-  orderNumber: string;
-  orderDate: string;
-  totalPrice: number;
-  paymentStatus: string;
-  deliveryAddress: Address;
-  deliveryDate: string;
-  status: string;
-  products: Product[];
-  createdAt: string;
-  userId: UserId;
-};
-
-export type UserId={
-  firstName: string;
-  lastName: string;
-  _id: string;
-}
+import { Orders} from '../../../custom.Types/orderTypes.ts';
 
 const AdminOrders: React.FC = () => {
   const [orders, setOrders] = useState<Orders[]>([]);
