+++ Frontend/src/custom.Types/userTypes.ts
@@ -5,3 +5,15 @@ export interface Address {
     postCode: string;
     country: string;
   };
+
+  export interface Member {
+    firstName: string;
+    lastName: string;
+    email: string;
+    address: Address;
+    membership: 'Basic' | 'Standard' | 'Premium' | 'Staff';
+    role: 'Member' | 'Admin';
+    is_activated: boolean;
+    _id: string;
+  }
+
